TEST: Implement test case, Upload photo wrong

+++ app/Http/Controllers/UserController.php
@@ -3,13 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Enums\UserPrefix;
+use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
+use App\Http\Requests\UploadProfilePictureRequest;
 use App\Models\User;
 use App\Traits\Pagination;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 use Inertia\Response;
+use Symfony\Component\HttpFoundation\Exception\BadRequestException;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
 class UserController extends Controller
 {
@@ -23,7 +28,7 @@ class UserController extends Controller
         $model = User::query()->rowNumber();
 
         return Inertia::render('Users/Users', [
-            'users' => $this->paginate($model, $request),
+            'users'        => $this->paginate($model, $request),
             'alertMessage' => session('alertMessage')
         ]);
     }
@@ -33,15 +38,24 @@ class UserController extends Controller
      */
     public function create()
     {
-        //
+        return Inertia::render('Users/CreateUser', [
+            'userPrefixes' => array_map(function ($prefix) {
+                return [
+                    'label' => $prefix,
+                    'value' => $prefix,
+                ];
+            }, array_column(UserPrefix::cases(), 'value'))
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(StoreUserRequest $request)
     {
-        //
+        User::create($request->validated());
+
+        return redirect()->route('users.index')->with('alertMessage', 'User has been created.');
     }
 
     /**
@@ -52,7 +66,7 @@ class UserController extends Controller
         $user = User::query()->findOrFail($id);
 
         return Inertia::render('Users/ViewUser', [
-            'user' => $user,
+            'user'         => $user,
             'userPrefixes' => array_map(function ($prefix) {
                 return [
                     'label' => $prefix,
@@ -70,7 +84,7 @@ class UserController extends Controller
         $user = User::query()->findOrFail($id);
 
         return Inertia::render('Users/EditUser', [
-            'user' => $user,
+            'user'         => $user,
             'userPrefixes' => array_map(function ($prefix) {
                 return [
                     'label' => $prefix,
@@ -101,8 +115,76 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
+        if ((int) $id === auth()->id()) {
+            throw new BadRequestException('You cannot delete yourself.');
+        }
+
         $model = User::destroy($id);
 
         return redirect()->route('users.index')->with('alertMessage', 'User has been deleted.');
+    }
+
+    /**
+     * Display a listing of the trashed resources.
+     *
+     * @param \Illuminate\Http\Request $request
+     *
+     * @return \Inertia\Response
+     */
+    public function trashed(Request $request): Response
+    {
+        $model = User::onlyTrashed()->rowNumber();
+
+        return Inertia::render('Users/TrashedUsers', [
+            'users'        => $this->paginate($model, $request),
+            'alertMessage' => session('alertMessage')
+        ]);
+    }
+
+    /**
+     * Restore the specified resource
+     *
+     * @param string $id
+     *
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function restore(string $id): RedirectResponse
+    {
+        $model = User::withTrashed()->findOrFail($id);
+        $model->restore();
+
+        return redirect()->route('users.index')->with('alertMessage', 'User has been restored.');
+    }
+
+    /**
+     * Permanently delete the specified resource
+     *
+     * @param string $id
+     *
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function delete(string $id): RedirectResponse
+    {
+        $model = User::withTrashed()->findOrFail($id);
+        $model->forceDelete();
+
+        return redirect()->route('users.index')->with('alertMessage', 'User has been permanently deleted.');
+    }
+
+    /**
+     * Upload a profile picture?
+     *
+     * @param \App\Http\Requests\UploadProfilePictureRequest $request
+     * @param string                                         $id
+     *
+     * @return RedirectResponse
+     */
+    public function upload(UploadProfilePictureRequest $request, string $id): RedirectResponse
+    {
+        $user = User::query()->findOrFail($id);
+
+        $user->update($request->validated());
+
+        return redirect()->route('users.edit', ['user' => $user]);
     }
 }
